Criação da validação de dados de entrada de uma Order, falta validar se as chaves estrangeiras são validas e existentes

+++ backend/src/orders/orders.controller.ts
@@ -1,8 +1,8 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, HttpException, HttpStatus } from '@nestjs/common';
 import { OrdersService } from './orders.service';
 import { CreateOrderDto } from './dto/create-order.dto';
 import { UpdateOrderDto } from './dto/update-order.dto';
-import { ApiTags } from '@nestjs/swagger';
+import { ApiResponse, ApiTags } from '@nestjs/swagger';
 
 @ApiTags('orders')
 @Controller('orders')
@@ -20,17 +20,56 @@ export class OrdersController {
   }
 
   @Get(':id')
+  @ApiResponse({ status: 404, description: 'Pedido não encontrado' })
   findOne(@Param('id') id: string) {
+
+    const order = this.ordersService.findOne(id);
+    if (!order) {
+      throw new HttpException(
+        'Pedido não encontrado',
+        HttpStatus.NOT_FOUND,
+      );
+    }
+
     return this.ordersService.findOne(id);
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateOrderDto: UpdateOrderDto) {
+  @ApiResponse({ status: 404, description: 'Pedido não encontrado' })
+  @ApiResponse({ status: 400, description: 'Data de pagamento inválida, o formato aceitável é: YYYY-MM-DD' })
+  async update(@Param('id') id: string, @Body() updateOrderDto: UpdateOrderDto) {
+
+    const order = this.ordersService.findOne(id);
+    if (!order) {
+      throw new HttpException(
+        'Pedido não encontrado',
+        HttpStatus.NOT_FOUND,
+      );
+    }
+
+    const orderPaymentMethod = updateOrderDto.payment_method;
+    if (orderPaymentMethod !== 'CREDIT_CARD' && orderPaymentMethod !== 'DEBIT_CARD' && orderPaymentMethod !== 'MONEY' && orderPaymentMethod !== 'PIX'){
+      throw new HttpException(
+        'Método de pagamento inválido, os metodos aceitáveis são: CREDIT_CARD, DEBIT_CARD, MONEY, PIX',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+    
     return this.ordersService.update(id, updateOrderDto);
   }
 
   @Delete(':id')
+  @ApiResponse({ status: 404, description: 'Pedido não encontrado' })
   remove(@Param('id') id: string) {
+
+    const order = this.ordersService.findOne(id);
+    if (!order) {
+      throw new HttpException(
+        'Pedido não encontrado',
+        HttpStatus.NOT_FOUND,
+      );
+    }
+
     return this.ordersService.remove(id);
   }
 }
